+++ plugins/prometheus/monitor.go
@@ -77,19 +77,30 @@ func (s *Plugin) Check(ctx context.Context, interval time.Duration) error {
 
 	// first check that the given queries have valid presets
 	for _, q := range s.config.Queries {
-		switch fmt.Sprintf("%s-%s", s.runtime, q.Preset) {
-		case "kubernetes-envoy-request-success":
-			querySQL = append(querySQL, KubernetesEnvoyRequestSuccess)
-		case "kubernetes-envoy-request-duration":
-			querySQL = append(querySQL, KubernetesEnvoyRequestDuration)
-		default:
-			return fmt.Errorf("preset query %s, does not exist", q.Preset)
+		if q.Preset != "" {
+			// use a preset if present
+			switch fmt.Sprintf("%s-%s", s.runtime, q.Preset) {
+			case "kubernetes-envoy-request-success":
+				querySQL = append(querySQL, KubernetesEnvoyRequestSuccess)
+			case "kubernetes-envoy-request-duration":
+				querySQL = append(querySQL, KubernetesEnvoyRequestDuration)
+			default:
+				return fmt.Errorf("preset query %s, does not exist", q.Preset)
+			}
+		} else {
+			// use the custom query
+			querySQL = append(querySQL, q.Query)
 		}
 	}
 
 	// execute the queries
 	for i, q := range querySQL {
 		query := s.config.Queries[i]
+
+		// check the query is not empty
+		if q == "" {
+			return fmt.Errorf("query %s is empty, please specify a valid Prometheus query", query.Name)
+		}
 
 		// add the interpolation for the queries
 		tmpl, err := template.New("query").Parse(q)